Add Value impl
